requirements.txt
Pillow
requests
icalendar
